+++ Explorers/Assets/_Scripts/Weapon/Bullet.cs
@@ -6,7 +6,7 @@ public class Bullet : MonoBehaviour
 {
     private Rigidbody _rb;
 
-    private int _damage;
+    private float _damage;
     private Vector3 _dir;
     private float _speed;
     private float _destoryTime;
@@ -37,7 +37,7 @@ public class Bullet : MonoBehaviour
         switch(other.tag)
         {
             case "Enemy":
-                other.GetComponent<Enemy>().TakeDamage(_damage);
+                other.GetComponent<Enemy>().TakeDamage((int)_damage);
                 Destroy(gameObject);
                 break;
             default:
